+++ .warp/bin/fix_help.sh
@@ -18,7 +18,8 @@ function fix_help_usage()
     warp_message_info   " --composer         $(warp_message 'fix permissions chmod() composer')"
     warp_message_info   " --addUser          $(warp_message 'add user docker and www-data to current user')"
     warp_message_info   " --owner            $(warp_message 'fix owner and read/write permissions inside the container')"
-    warp_message_info   " --fast             $(warp_message 'add group www-data && ug+rw permissions inside the container')"
+    warp_message_info   " --fast             $(warp_message 'add group www-data and ug+rw permissions inside the container')"
+    warp_message_info   " --sandbox          $(warp_message 'Fix owner, ug+rw permissions and turns www-data user into /var/www/html/<version>-ce/bin/magento proprietary')"
     warp_message ""
 
     warp_message ""